+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * New stat: customers with most no-shows
+* Bookings list: click on customer name to display list of all bookings
 
 ### 1.7.1: July 7th, 2021
 * Store Aquos error in booking entry as _aquos_status